Have implemented varying color and have 2 Mesh instances appearing in Game cpp. Working on getting GameObject working properly.

+++ Framework/Source/Events/Event.cpp
@@ -22,7 +22,7 @@ namespace fw {
     InputEvent::InputEvent(DeviceType dType, InputState iState, WPARAM wParam) :
         m_DeviceType(dType),
         m_InputState(iState),
-        m_ID(wParam)
+        m_ID((int)wParam)
     {
 
     }
